+++ src/cleanScreen.ts
@@ -1,0 +1,3 @@
+export function cleanScreen() {
+  process.stdout.write('\x1Bc')
+}